Split a blender class to new file
