+++ opencog/python/blending/src/base_blender.py
@@ -1,0 +1,13 @@
+from abc import ABCMeta, abstractmethod
+
+__author__ = 'DongMin Kim'
+
+class Blender(object):
+    __metaclass__ = ABCMeta
+
+    def __init__(self, atomspace):
+        self.a = atomspace
+
+    @abstractmethod
+    def blend(self):
+        pass
